+++ src/headers/DelayImportDir.cpp
@@ -6,8 +6,8 @@
 
 using namespace Pepper;
 
-int DelayImportDir::s_diskToMemDiff = 0;
-template<> int *DelayImportDescriptor::s_pDiskToMemDiff = &DelayImportDir::s_diskToMemDiff;
+size_t DelayImportDir::s_diskToMemDiff = 0;
+template<> size_t *DelayImportDescriptor::s_pDiskToMemDiff = &DelayImportDir::s_diskToMemDiff;
 
 DelayImportDir::DelayImportDir(const PeFile &pe, const FileBytes &fbytes, const DataDirectoryEntry &dde)
 : IDirectory(pe, fbytes, dde)
@@ -17,9 +17,9 @@ DelayImportDir::DelayImportDir(const PeFile &pe, const FileBytes &fbytes, const 
         // append non-null IMPORT_DESCRIPTOR to m_descriptors vector
         char null[sizeof(IMAGE_DELAY_IMPORT_DESCRIPTOR)] = {0};
         IMAGE_DELAY_IMPORT_DESCRIPTOR *pid = (PIMAGE_DELAY_IMPORT_DESCRIPTOR)dir();
-        int i=0;
+        size_t i=0;
         while (memcmp(&pid[i], &null, sizeof(null))) {
-            m_descriptors.emplace_back(pe, fbytes, dirOffset() + (i * (int)sizeof(null)));
+            m_descriptors.emplace_back(pe, fbytes, dirOffset() + (i * sizeof(null)));
             i++;
         }
     }
